pyproject.toml
[tool.poetry]
name = "localllminspect"
version = "0.1.0"
description = "Local LLM Inspector - OpenAI-compatible API proxy with LLM trace visualization"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.0"
langfuse = "^2.57.7"
openai = "^1.12.0"


[tool.poetry.dev-dependencies]
black = "^23.0"
flake8 = "^6.0"
mypy = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"